Refactor import statements to use src module structure.
